Move util.cpp/util.hpp to libipmid
These are functions that are used widely by ipmid providers, so it makes
sense to put them in libipmi.so (the library that all providers must
link against).

Tested-by: use nm to inspect the binaries to see that the symbols are in
           the expected library.

           arm-openbmc-linux-gnueabi-nm libipmid.so.0.0.0 \
                      | grep getDbusObject
           0001063c T _ZN4ipmi13getDbusObjectERN9sdbusplus....
