+++ apphandler.cpp
@@ -6,7 +6,6 @@
 #include "sys_info_param.hpp"
 #include "transporthandler.hpp"
 #include "types.hpp"
-#include "utils.hpp"
 
 #include <arpa/inet.h>
 #include <ipmid/api.h>
@@ -19,6 +18,7 @@
 #include <array>
 #include <cstddef>
 #include <fstream>
+#include <ipmid/utils.hpp>
 #include <memory>
 #include <nlohmann/json.hpp>
 #include <phosphor-logging/elog-errors.hpp>
