+++ chassishandler.cpp
@@ -5,7 +5,6 @@
 #include "ipmid.hpp"
 #include "settings.hpp"
 #include "types.hpp"
-#include "utils.hpp"
 
 #include <arpa/inet.h>
 #include <endian.h>
@@ -19,6 +18,7 @@
 #include <cstring>
 #include <fstream>
 #include <future>
+#include <ipmid/utils.hpp>
 #include <map>
 #include <phosphor-logging/elog-errors.hpp>
 #include <phosphor-logging/log.hpp>
