+++ globalhandler.cpp
@@ -1,9 +1,8 @@
 #include "globalhandler.hpp"
 
-#include "utils.hpp"
-
 #include <ipmid/api.h>
 
+#include <ipmid/utils.hpp>
 #include <phosphor-logging/elog-errors.hpp>
 #include <phosphor-logging/log.hpp>
 #include <string>
