+++ include/ipmid/utils.hpp
